new: container prune
