+++ lib/l10n/app_ja.arb
@@ -132,6 +132,7 @@
   "preview": "プレビュー",
   "privateKey": "秘密鍵",
   "process": "プロセス",
+  "prune": "剪定する",
   "pushToken": "プッシュトークン",
   "pveIgnoreCertTip": "オプションを有効にすることは推奨されません、セキュリティリスクに注意してください！PVEのデフォルト証明書を使用している場合は、このオプションを有効にする必要があります。",
   "pveLoginFailed": "ログインに失敗しました。Linux PAMログインのためにサーバー構成からのユーザー名/パスワードで認証できません。",
