+++ lib/l10n/app_pt.arb
@@ -132,6 +132,7 @@
   "preview": "Pré-visualização",
   "privateKey": "Chave privada",
   "process": "Processo",
+  "prune": "Podar",
   "pushToken": "Token de notificação push",
   "pveIgnoreCertTip": "Não recomendado para ativar, cuidado com os riscos de segurança! Se estiver usando o certificado padrão do PVE, você precisa habilitar esta opção.",
   "pveLoginFailed": "Falha no login. Não é possível autenticar com o nome de usuário/senha da configuração do servidor para login no Linux PAM.",
