+++ lib/l10n/app_ru.arb
@@ -132,6 +132,7 @@
   "preview": "Предпросмотр",
   "privateKey": "Приватный ключ",
   "process": "Процесс",
+  "prune": "Обрезать",
   "pushToken": "Токен уведомлений",
   "pveIgnoreCertTip": "Не рекомендуется включать, обратите внимание на риски безопасности! Если вы используете стандартный сертификат от PVE, вам нужно включить эту опцию.",
   "pveLoginFailed": "Ошибка входа. Невозможно аутентифицироваться с помощью имени пользователя/пароля из конфигурации сервера для входа в Linux PAM.",
