+++ lib/l10n/app_uk.arb
@@ -132,6 +132,7 @@
   "preview": "Попередній перегляд",
   "privateKey": "Приватний ключ",
   "process": "Процес",
+  "prune": "Обрізати",
   "pushToken": "Надіслати токен",
   "pveIgnoreCertTip": "Не рекомендується включати, будьте обережні з ризиками безпеки! Якщо ви використовуєте стандартний сертифікат від PVE, вам потрібно увімкнути цю опцію.",
   "pveLoginFailed": "Не вдалося увійти. Неможливо пройти аутентифікацію за допомогою імені користувача/пароля з конфігурації сервера для входу Linux PAM.",
