+++ lib/l10n/app_zh_tw.arb
@@ -132,6 +132,7 @@
   "preview": "預覽",
   "privateKey": "私鑰",
   "process": "行程",
+  "prune": "修剪",
   "pushToken": "消息推送 Token",
   "pveIgnoreCertTip": "不建議啟用，請注意安全風險！如果您使用的是 PVE 的默認證書，則需要啟用此選項。",
   "pveLoginFailed": "登錄失敗。無法使用伺服器配置中的使用者名稱/密碼以 Linux PAM 方式登錄。",
